Use existing tooltip component for the modal
