+++ app/modules/tooltips/_styles.scss
@@ -98,12 +98,12 @@ $mt2_arrowheight: .625rem;
 
 // Tooltip Modal
 
-.mt2_tooltip-modal-btn {
+.mt2_tooltip-modalbtn {
   cursor: pointer;
   padding: 1px 6px 0;
 }
 
-.mt2_tooltip-modal-content {
+.mt2_tooltip-modalcontent {
   left: 50%;
   position: absolute;
   top: 50%;
