+++ app/modules/tooltips/docs.html
@@ -61,10 +61,11 @@
 
     <p class="mt2_subh4">This tooltip opens a modal</p>
 
-    <div data-pestle-module="TooltipModal">
+    <div data-pestle-module="Tooltip">
       <script type="text/json" data-pestle-options>
         {
-          "tooltipContent" : "Hello World, inside a modal!"
+          "tooltipContent" : "Hello World, inside a modal!",
+          "modal" : true
         }
       </script>
     </div>
