+++ app/modules/tooltips/scripts/MTTooltip.jsx
@@ -3,8 +3,42 @@
 import React, { Component, PropTypes }  from 'react';
 import {Pestle} from '@natgeo/mortar-pestle';
 import classNames from 'classnames';
+import {SimpleModal} from './modal.jsx';
+
 
 class Tooltip extends Component {
+  static propTypes = {
+    tooltipContent: PropTypes.string,
+    placement: PropTypes.oneOf(['top', 'left', 'right', 'bottom']),
+    modal: PropTypes.bool
+  };
+
+  static defaultProps = {
+    placement: 'top',
+    modal: false
+  };
+
+  render(){
+    if(this.props.modal){
+      return <TooltipModal {...this.props} />
+    }
+    else {
+      return <TooltipBubble {...this.props} />
+    }
+  }
+}
+
+
+class TooltipBubble extends Component{
+  static propTypes = {
+    tooltipContent: PropTypes.string,
+    placement: PropTypes.oneOf(['top', 'left', 'right', 'bottom'])
+  };
+
+  static defaultProps = {
+    placement: 'top'
+  };
+
   constructor(props) {
     super(props);
 
@@ -55,10 +89,6 @@ class Tooltip extends Component {
       'mt2_tooltip--right': this.props.placement === 'right'
     });
 
-    if(tooltipPlacement === ''){
-      tooltipPlacement = 'mt2_tooltip--top';
-    }
-
     return(
       <div ref="tooltipContainer" className={`mt2_tooltip mt2_subh4 mt2_bgcolor--neutral--xxl ${tooltipPlacement}`}>
         <p dangerouslySetInnerHTML={{__html: this.props.tooltipContent}} />
@@ -78,9 +108,51 @@ class Tooltip extends Component {
   }
 }
 
-Tooltip.propTypes = {
-  tooltipContent: PropTypes.string,
-  placement: PropTypes.oneOf(['top', 'left', 'right', 'bottom'])
+class TooltipModal extends Component {
+  static propTypes = {
+    tooltipContent: PropTypes.string
+  };
+
+  constructor(props){
+    super(props);
+    this.state = {
+      active: false
+    };
+
+    this.open = this.open.bind(this);
+    this.close = this.close.bind(this);
+  }
+  open(){
+    this.setState({active: true});
+  }
+  close(){
+    this.setState({active: false});
+  }
+  renderModal(){
+    const tooltipContent = this.props.tooltipContent || '';
+
+    return (
+      <SimpleModal onClose={this.close}>
+        <div className="mt2_tooltip-modalcontent">{tooltipContent}</div>
+        {this.props.children}
+      </SimpleModal>
+    )
+  }
+  render(){
+    const buttonClasses = `
+      mt2_tooltip-modalbtn
+      mt2_h5 mt2_color--neutral--xd
+      mt2_bordercolor--neutral--xl
+      mt2_bgcolor--neutral--xxxl`;
+
+    return (
+      <div>
+        <button className={buttonClasses} onClick={this.open}>?</button>
+        {this.state.active ? this.renderModal() : ''}
+      </div>
+    );
+  }
 }
+
 
 export default Tooltip;
